+++ legacy/src/utils/codes-lists/codes-lists-utils.ts
@@ -1,3 +1,5 @@
+import { CodeList, RemoteCodeList } from '@/models/code-lists';
+
 /**
  * This method will check if a code list contains a given child
  */
@@ -8,3 +10,13 @@ export function hasChild(
   const found = listCodes.find((listCode) => listCode.parent === code.value);
   return found !== undefined;
 }
+
+const isNullish = (param: unknown) => param === undefined || param === null;
+
+export const isNomenclatureCodeList = (codeList?: CodeList): boolean =>
+  !isNullish(codeList?.urn) || !isNullish(codeList?.suggesterParameters);
+
+export const isNomenclatureRemoteCodeList = (
+  codeList?: RemoteCodeList,
+): boolean =>
+  !isNullish(codeList?.Urn) || !isNullish(codeList?.SuggesterParameters);